NP-2717 Refactor Vocabulary components
