+++ src/pages/registration/description_tab/vocabularies/HrcsCategoryInput.tsx
@@ -1,0 +1,16 @@
+import { useTranslation } from 'react-i18next';
+import { VocabularyAutocomplete, VocabularyComponentProps } from './VocabularyAutocomplete';
+import { hrcsCategories } from '../../../../resources/vocabularies/hrcsCategories';
+
+export const HrcsCategoryInput = (props: VocabularyComponentProps) => {
+  const { t } = useTranslation('registration');
+
+  return (
+    <VocabularyAutocomplete
+      {...props}
+      options={hrcsCategories.categories}
+      id="hrcs-categories"
+      label={t('description.hrcs_categories')}
+    />
+  );
+};
